refactor(executor): Split fast_path.rs into modular directory structure
Splits the 2,279-line fast_path.rs into a well-organized module directory
following existing codebase patterns (nonagg/, columnar_execution/).

Module structure:
- fast_path/mod.rs: Entry point and execute_fast_path method (~100 lines)
- fast_path/analysis.rs: Query analysis functions (~400 lines)
- fast_path/helpers.rs: Shared utilities (~350 lines)
- fast_path/pk_lookup.rs: Primary key lookup strategies (~250 lines)
- fast_path/index_lookup.rs: Secondary index lookups (~350 lines)
- fast_path/range_scan.rs: PK range scan with early projection (~170 lines)
- fast_path/streaming_agg.rs: Streaming aggregation (~120 lines)

Benefits:
- Query analysis is now isolated and independently testable
- Each execution strategy is in its own module
- All existing public APIs remain accessible
- No performance regression (fast path remains fast)
- All 1861 tests continue to pass

Closes #4123
